requirements.txt
python-telegram-bot==13.7
requests==2.28.1
feedparser==6.0.10
python-dateutil==2.8.2
beautifulsoup4==4.11.1